refactor: 도메인 -> 출력용 문자열 변환 로직을 Dto에서 Mapper클래스로 이관
